+++ backend/src/main/java/it/autoflow/configuration/entity/OptionalAccessorio.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 
 @Getter @Setter
 @Entity
@@ -24,5 +23,5 @@ public class OptionalAccessorio {
     private String descrizione;
 
     @Column(nullable = false)
-    private BigDecimal prezzo;
+    private Double prezzo;
 }
