+++ backend/src/main/java/it/autoflow/invoice/entity/Fattura.java
@@ -7,7 +7,6 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter @Setter
@@ -32,7 +31,7 @@ public class Fattura {
     private Proposta proposta;
 
     @Column(nullable = false)
-    private BigDecimal importoTotale;
+    private Double importoTotale;
 
     private LocalDate dataPagamento;
